Improvement: change handle position for context menu
Since deltachat supports custom backgrounds, the handle for context
menus (only visible on hover) should not be near the bubble but in
the upper right corner. Otherwise it might not have enough contrast.

I added 10px padding-top to the text to enable a proper distance,
may be there is a better solution...