use less muted color for "recently seen"
with the "connection view",
we already have a green indicating sth. as "online", reuse that.
the green used before is a bit too much of a muted color,
this does not reflect "activity" well.

the difference is not huge, however,
this pr also synchronizes the color with android and ios
and also comparable indicators
in other apps have less muted colors for showing activity.

cmp https://github.com/deltachat/deltachat-android/pull/2393
and https://github.com/deltachat/deltachat-ios/pull/1707