show at least a basic hint on "Welcome / Add Second Device"
the idea of the extra "Welcome / Add Second Device" button was to guide users,
that install Delta Chat on a second device without even being aware
of the settings option on the other device.

without the hint to the other device,
the user has no idea, how to set up a second device
or what to scan.

the dialog may be improved eg. also showing
`multidevice_same_network_hint` and `multidevice_experimental_hint`,
however, that does not fit to the existing dialog -
therefore, this pr fixes the very basic things only.