streamline headlines in profiles
the headlines in profiles should just read what the "profile" is about -
so "Contact", "Group", "Broadcast List" -
not that it is a "profile" (whatever that is from the view of the user).

also the 'Edit' prefix is not correct,
esp. confusing as there is a dialog that open the thing for editing.

these are the headlines that we're using also on android/ios,
and that are used in similar ways also in similar app.