first draft
still has manual play ui, that would need some design work to look good.
I still like that approach more (disconnect preview and set action),
but I'll for now go with the select on play method to get the others to agree to merge it.
That approach will need an extra "preview notification" button, which is a good thing anyway outside of this topic, to test if the notification system works.