Use native watch mode and CSS modules of esbuild (#3571)
This PR introduces CSS Modules into our build pipeline. It's actually
natively supported by esbuild, I just needed to update it to the latest
version and add TypeScript support via a plugin. There's a bit of noise
in this PR as I needed to bump the typescript and all plugins, so
prettier got sad and I needed to autofix that, but most of the touched
files can be ignored during review as they only changed formatting.

On top this uses a thin wrapper around the css-loader of esbuild and
extend it with SCSS functionality, so we don't need to rewrite our .scss
files to pure CSS.

To still get linter warnings I needed to add a eslint plugin and a
reporter plugin which tells us about the current bundle process. On top
when starting the npm run watch script, a parallel tsc process is
launched in the background.

The idea behind this PR is to start using (S)CSS modules for new
components and refactor old components to use them step by step. It will
hopefully help us to structure the CSS better by avoiding global scope,
also it will be easier to see where the styles for each component live.

* Update esbuild to support css modules

* Allow scss in style modules

* Include the generated css module bundle as well

* Add plugin to get types from css modules

* Move css module ts plugin into shared tsconfig.json

* Extend prettier to check .scss files everywhere

* Make prettier happy

* Add entry to CHANGELOG.md

* Use native esbuild watch process, add eslint reporter

* Add doc-strings

* Update CHANGELOG.md

* Remove redundant logging

* Plugin needs to be used in all tsconfig.json files

* Update all typescript and tools around it

* Start a tsc process in the background

* Update prettier plugin after update

* Make prettier happy

* Start watching process in npm script instead

* Use types explicitly for Node v18

* Improve CHANGELOG.md message

* Do not run eslint on builds

* Also run eslint on production builds

* Remove unnecessary plugin from tsconfig.json files

* Fix types for removing falsy context menu items

* Better naming

* Make type name singular