fix: i18n: wrong pluralization for many languages
Closes #3870

Also fix some strings being displayed as keys and not actual strings,
e.g. "1 minute" message age in Indonesian, which would show as
"n_minutes:one" instead of "1 menit".

However, some strings are still not pluralized at all, such as
`autodel_device_ask`, but let's consider that a different issue.

This is somewhat of an alternative to #3889 as a fix to the
"message age" part of #3870