refactor: safer types
To avoid issues like #3990

This ensures that the type and the array always have the same keys

However, simply adding an extra property on `SettingsValues`
will not cause errors, it will simply keep
the `SettingsStoreState` type unchanged.
But any runtime changes (e.g. adding or accessing a property
without also adding it to the array, or vice versa) will be caught.