fix: chat "scrolling up" upon reaction
Closes #3753

This removes `scrollToBottom` JS on each new message,
which is now unnecessary. However, the old code used `ifClose: true`,
which would still scroll the chat if it was scrolled up less than
7 pixels up. The new implementation still has similar behavior as
the new scroll anchor is not at the very bottom
of the scrollable area, so the chat will still show the new message
as long as the bottom edge of the previous oldest message
was still visible when the new one arrived.