feat: paste invite link to search to add contact
Adds a pseudo item similar to "add contact" when you type an email

The "standard" way of pasting the invite link under the gear button
on the second tab of the "scan QR" dialog is not very straigtforward.

The first reflex is to paste it in the search field,
so let's handle this.