fix: chatlist image thumbnails not showing on Win
The `url()` wants a string, so let's give it a string.
Otherwise it treats backwards slashes, which used in Windows paths,
as escape characters, which results in an invalid URL.

The bug was introduced in 72bb581da4,
then there was a fix for something in 87ba793e66d.

Though it feels like there should be a proper way to convert file
paths to URL...