style: cancel old message highlight animations
I.e. when we highlight a new message as a result of `jumpToMessage`,
remove the animation on the previously highlighted message
(or messages).

This is particularly useful for the Ctrl + Up shortcut
which is to choose the message to reply to.

Just FYI the added code usually takes 0.1-0.5ms to execute.