improvement: a11y: focusable contact list items
This is quite a significant change.
I tested a lot of placed where the affected code is used,
but still might have missed some regression.

And perhaps the behavior is still not ideal in some places,
but it's better to assess it in its final form after we add
[the roving tabindex functionality (arrow key accessibility)](https://github.com/deltachat/deltachat-desktop/pull/4224).
Either way I think this is an improvement.