improvement: better scroll for jumpToMessage
Add `scrollIntoViewArg` parameter to `jumpToMessage`
and specify different values for it depending on context.

More specifically, prefer scrolling the target message
to the center, e.g. when showing the first unread message.
Prior to this commit, we'd scroll the first unread message
such that it's at the bottom of the message list,
which can be considered a regression, which was introduced in
5f0efe169a56113f25d6aee48fab045956c83a00.

Closes https://github.com/deltachat/deltachat-desktop/issues/4284.