remove unused `runtime.resolveThemeAddress`.
the result was not used so I thought maybe the error was used? but turns out it called the wrong electron ipc function. And I can't think of an important reason why it should be fixed instead of removed. we don't store absolute paths for themes in config, maybe we did in the past, but now we certainly don't need it anymore