add theme support in browser edition (#4304)
* remove unused `runtime.resolveThemeAddress`.

the result was not used so I thought maybe the error was used? but turns out it called the wrong electron ipc function. And I can't think of an important reason why it should be fixed instead of removed. we don't store absolute paths for themes in config, maybe we did in the past, but now we certainly don't need it anymore

* sort event callback functions by whether we can implement them

* make themes work in browser edition

* aff doc comment to runtime.onThemeUpdate

* add changelog entry