fix: potential crash due to `Component({ ...props`
We already had similar fixes in the past.
I have looked through the code and have found no more of such calls.

Without this we get an error in the console after pressing "OK"
in the "Add Members" dialog in "Create Group":
`Warning: React has detected a change in the order of Hooks called by ContactList`.