fix: chat not being scrolled to bottom (images)
Closes https://github.com/deltachat/deltachat-desktop/issues/4404.

The bug was introduced in db8327931ef.
Prior to that all attachments had a fixed height of 200px.

This should also eliminate layout shifting when scrolling the chat,
and when new messages arrive.

An alternative solution that would still have layout shifting,
but at least keep the chat "anchored" to bottom would be
https://github.com/deltachat/deltachat-desktop/issues/3753#issuecomment-2336809190,
but it's not supported well by browsers.

Co-authored-by: Max Philippov <maxphilippov@users.noreply.github.com>