fix: chat "scrolling up" because of imgs loading
This, together with https://github.com/deltachat/deltachat-desktop/pull/4407,
should close https://github.com/deltachat/deltachat-desktop/issues/4404.

The bug was introduced in https://github.com/deltachat/deltachat-desktop/pull/3999.
Pior to that we used to have fixed width on quoted images as well.
This commmit practically reverts that MR, but also adds
`object-fit` and `object-position`, and is now in line with
message attachment images.