style: stickers being smaller than needed
Because of `max-width: 80%`, which refers to 80% of the wrapper
square, and not the chat section
(the later was apparently the intention).
This has been introduced in https://github.com/deltachat/deltachat-desktop/pull/4315,
which has later been sort of reverted, but not completely, in
https://github.com/deltachat/deltachat-desktop/pull/4407.
So, let's remove the residuals, especially that stickers also have
fixed `height: 200px`, so `max-height` doesn't make sense.