refactor: unify "Select Chat": use one component
For:
- webxdc `sendToChat()`
- mailto dialog
- "Forward Message" dialog

What this commit comes down to:
- Take the ForwardTo dialog and use it to create
  a more generic, reusable SelectChat dialog.
- Make those similar dialogs use the new SelectChat component,
  instead of defining their markup and logic manually.

The dialogs have practically the same markup and the same
search logic, with very little differences.
They apparently were copy-pasted from `ForwardMessage` dialog,
which is the oldest of them all.

It's getting pretty annoying to maintain them, with all the
`RovingTabindex` and `InfiniteLoader`, so, it's about time
to unify them.

This is mainly a refactor, but causes a few small visual changes.
For example, this fixes a small visual bug
in the `WebxdcSendToChat` dialog
where the search field takes full height
when there are no search results.