fix: `jumpToMessage` not working sometimes
...when the chatId of the message is different
from the currently open chat.

Closes https://github.com/deltachat/deltachat-desktop/issues/4508.
This is probably related to the React upgrade
(https://github.com/deltachat/deltachat-desktop/pull/3958),
which changed the behaviour of `useEffect`.

The issue was that we called `window.__internal_jump_to_message`
_before_ `MessageListStore` has been reinstantiated
after the change of chatId (in `useMessageList`),
so we'd call it on an old instance.

I think there might still be cases where this bug appears,
but at least this makes it appear less often.