refactor: fix message list potentially being empty
...sometimes.
And maybe sometimes showing messages from previous chat.

I have not been able to confirm whether this fixes
an actual bug.
But at least in my understanding this fixes a semantic error.

The presence of the error is obvious if you make
`store.effect.loadChat()` synchronous.
So, what could be happening is that `loadChat` finishes _before_
we subscribe to the new store instance,
and thus the state never gets updated,
at least not until the next store update, which could be e.g.
`fetchMoreMessagesTop`.