fix: jumpToMessage flashing the bottom of new chat
That is, when you jump to a message in a different chat
(e.g. by clicking on a notification, or from message search,
or from a "private reply" quote),
we would momentarily show the new chat scrolled to bottom
prior to jumping to the desired message.

Why this is bad:

- We mark the visible messages as read.
  In this case that would be the last messages in the chat.
  But they weren't actually read, the user did not intend
  to scroll to them.
- Extra layout shifting: not great visually.
- Bad for performance. Let's just load the part of the chat
  that we need right away.

Related:
- https://github.com/deltachat/deltachat-desktop/issues/4508
- https://github.com/deltachat/deltachat-desktop/pull/4510
- https://github.com/deltachat/deltachat-desktop/pull/4554