improvement: don't clear search on focus
Clearing the input is very easy anyway.
We should not clear it automatically in case e.g. the user wants
to come back to it and adjust the search,
or if they're simply using it as a convenient way to get the focus
closer to when they need it to be.