feat: display past chat members
Since the core now has a new group consistency algorithm, the list of past members who left the group in the last 90 days is sent around in the (encrypted) header. Displaying it makes users aware that this happens and past members are known even to the new members. Besides "security" aspect, the feature is just occasionally useful to check "have I removed the bot from the group already, or just cannot find it in the huge member list?".

For use together with deltachat/deltachat-core-rust#6442

Co-authored-by: WofWca <wofwca@protonmail.com>