fix: a11y: improve tab order (navbar)
This ensures that the list of chats follows the search bar,
and that the messages list follows the chat header.
Closes https://github.com/deltachat/deltachat-desktop/issues/4225.

This mostly moves code around, and simplifies some things
(such as replacing `width: 30%` duplication) along the way.

But it also changes the relative sizes of the "chat list" and the
"chat" (main view) sections, to a ratio of 3:8
(before the "chat list" section was 30% of the window size).

Otherwise, besides, of course, the tab order fix,
this is not supposed to affect the behavior and the look of the app.