move "Save" and "Save as" together
> this PR is the outcome of some one to one discussions

even though "Save" saves to saved messages
and "Save As" saves a file to some other place,
user may not think much about that,
and it is more confusing if they are not listed together as usual.

background: there are other ideas around,
eg. to reword "Saved Messages" to "Bookmarks"
(which would have the disadvantage to change a known UI element,
and it would not fit that nicely to adding things manually,
and it would not reflect the presistence).
another idea is to use "Export" instead,
however that was used before  https://github.com/deltachat/deltachat-desktop/pull/2546/
and probably cahnged for $reasons
(also cmp. https://github.com/deltachat/deltachat-desktop/pull/2160 )
so, all in all, the simple moving of the option seems to be a good first approach,
we can iterate if ppl at scale are confused (and not only us devs :)