refactor: HTML msg viewer: replace `windowId` arg
In Tauri, window labels are important for permission control,
and, prior to this commit, `open_html_window` used the `window_id`
string parameter for making the label
of the respective HTML message viewer window.
This commit limits the amount of control that the caller gets
over the window label, by replacing `window_id` string
with just two numbers.
So, e.g. now it's not possible to inject things like "main" there.

I have tested this change on both Electron and Tauri.