improve toast after copying link to clipboard
the dialog for copying invite link does not say anything about a "QR code",
and the old text saying "Copied QR url to clipboard" opens additional questions -
what is a "QR url"?

this is esp. bad as the toast disappears while thinking about it :)

so, better just use the default text "Copied to clipboard" - the toast
is anyway more a visual feedback that sth. has happend, not so much sth. to read.

came over that as translators have issues with the existing "QR url" text.