tauri: hardening: invalid proxy for network isolation
Making a browser use a proxy is arguably the most powerful way
(out of all feasible ways) to disable internet access
for webxdc apps.
Let's use it!

I have verified that on Windows (Chromium) the proxy still
doesn't receive any packets (because of `host-rules`).
I also tried disabling `host-rules`, in which case the packets
to the dummy proxy do get sent, but I was unable to produce
any kind of other network activity, not DNS prefetch, not WebRTC,
even if I removed the webxdc CSP.