tauri: fix: deny navigation to https://webxdc.localhost
The `on_navigation` didn't fully check the URL origin
on Windows and Android, so a webxdc app could navigate
to `https://webxdc.localhost`, which could potentially
produce network traffic on `localhost:443`.

Fortunately, thanks to the presence of `host-rules`
and `proxy_url`, this wouldn't happen
even if navigation were to take place,
but let's still properly deny it.