fix: a11y: correct `aria-posinset` for chat list
Due to the usage of a virtualized list,
screen readers would announce an incorrect number of chats
in the chat list, and the positions of individual chats in the list.

Explicitly setting `aria-posinset` and `aria-setsize` should work.

This only handles the chat list and not all the virtualized lists
that we have.

Related:
- https://github.com/bvaughn/react-window/issues/808.
- https://github.com/deltachat/deltachat-desktop/pull/4660.
- https://github.com/deltachat/deltachat-desktop/pull/5025.