fix: a11y: unmark "Chats" section header as <nav>
The "Chats" and the "Messages" sections headers used to be
one single element, then it has been split into two in
66c96d7df19f591a8c8610842952cb9f7b9ed540
(https://github.com/deltachat/deltachat-desktop/pull/4672).
But then it started to make even less sense for it to be a <nav>.

It basically consists of only two elements,
so it's probably not worth having a separate landmark for it.
Let's turn it into a generic <section> with no label.