improvement: a11y: "View Profile" label for header
This is foremost a preparation for making the chat heading
an h1-h6 element, so that it can act as a section header.
Putting an h1-h6 element inside of a button is not allowed,
and putting a button inside of a h1-h6 element is a bit awkward,
so let's have a separate button for "View Profile" / "Edit Group".