fix: webxdc: same `selfAddr` on different webxdcs (#5068)
Closes https://github.com/deltachat/deltachat-desktop/issues/4993,
and likely https://github.com/deltachat/deltachat-desktop/issues/4401.

The issue is that we register the `webxdc:` scheme protocol handler
once per account, when you open the first webxdc, and when we do,
we incorrectly use the captured `webxdcInfo` var inside the handler,
resulting in us using `selfAddr` of the first opened webxdc app
for all the subsequently opened apps.

The fix is to get `selfAddr` from the respective webxdcInfo object
for each app instance.
This commit also does the same for other `webxdcInfo` properties,
such as `sendUpdateInterval`.

`selfAddr` used to be the email address, so this used to be fine,
but then `selfAddr` started to be different between messages
(https://github.com/deltachat/deltachat-desktop/issues/4401).