tauri: hardening: sending msg on `dcnotification:`
This does not fix an exploitable bug, but is instead only a refactor.
The reason this is not exploitable is because
`decode_deeplink` always returns `user_text: None`,
thus making the `send_reply` path unreachable.

The threat is that a malicious website could create and `.click()`
a `dcnotification:` link with "Reply" payload.
A (good) browser would show a confirmation dialog
saying something along the lines of
"example.com wants to open deltachat-tauri. Allow?".
If the user clicked "allow", Delta Chat would handle the link
and send a message automatically.