feat: play sound on incoming msg in current chat
This is mostly to address an accessibility issue
where screen reader users are not notified of incoming messages.

Applying `aria-live` to the MessageList and thus announcing
every message's text is something that requires
a lot of consideration.
So let's settle on this simple solution for now instead.
When you hear the sound, you know that it's time to focus
the message list to go through the messages.

See
- https://github.com/deltachat/deltachat-desktop/issues/4743.
- https://github.com/deltachat/deltachat-desktop/pull/4692.

The new setting to control the volume of the sound is also present
on Delta Chat Android, except that it's only an on-off toggle.