improvement: reverse recent webxdcs order
See https://github.com/deltachat/deltachat-desktop/pull/5151#issuecomment-2909576653.
The argument for the original order (with `mediaIds.reverse()`)
was that the order matches the order in the gallery dialog.
However, I don't think that this matching is important.
What's more important is that it's more intuitive that
the items that is the last one (the rightmost) corresponds to the
message that was sent the last.
Otherwise I personally need to think really really hard
which app I'm about to open (I have several instances
of the same app).