refactor: simplify ChatContext `selectChat`
Use the more declarative `useRpcFetch` instead of manual
race condition mitigation.

This also surfaced the potential bug with setting
`__internal_jump_to_message_asap` too late,
which resulted in the initial load of the message list
showing the last message of the chat (the default view)
instead of jumping right to the target message.
So I fixed it.

I have tested the change for a while.