fix: a11y: delivery status being announced twice
The issue was that apparently NVDA first announces the label
(`aria-label`) of the changed element, and then announces
the change itself, which, in this case, is a change to
`aria-label`. This results in it simply announcing the label twice.

Let's mitigate this by removing `aria-label`.
And along the way, let's split the button
into a `role='status'` element and a "Message Info" button.