fix: allow offline usage of app picker
Don't show the list of apps only if we failed to fetch it,
and not whenever we're offline.

This also fixes an annoying bug where the app picker
would show "offline" if you opened it
during "updating" network status,
which notably also happens when you just focused the window
(due to a `rpc.maybeNetwork()` call).
And this has been causing E2E test failures recently.

`getHttpResponse` is able to cache data since not so long ago,
so let's utilize that.

Note that the fact that we have cached the list of apps
doesn't mean that all the apps themselved are cached,
so users might still encounter an error.

However, if the list of apps and the app file is cached,
this allows you, even when offline,
to send an app to the chat and open it.

This change also makes sense even if `getHttpResponse`
didn't do any caching. We do not rely on this.