fix: remove RTCPeerConnection exhaustion (FILL500)
This partially reverts 750721ca5b8e4e9c1c376c95cec470eefefd474c
(https://github.com/deltachat/deltachat-desktop/pull/3098).
Not completely though: the `<iframe>` thing is still there,
and I have not removed it yet for the sake of being conservative.

The most important point of this change is to prepare ourselves
for the removal of the `IsolateSandboxedIframes` Chromium feature.
It has been enabled by default, and is to be removed soon-ish:
https://issues.chromium.org/issues/362246896.
The feature makes the RTCPeerConnection exhaustion hack
not work anymore, because an attacker can create
a sandboxed `<iframe>`, and, as the feature name implies,
process-isolated, thus have a separate RTCPeerConnection 500 pool.

Earlier we simply un-enabled this Chromium feature in
50bcd88a80cfb6a86e30efae4f723ab46f3d08aa
(https://github.com/deltachat/deltachat-desktop/pull/4351),
but when it's removed, we will not be able to.
We could hold off on upgrading Electron then, but let's not.

To answer "Why haven't we done this before?":
probably because the 500 hack was the first thing
that proved to work, both on Electron and in Android.
But in Android apparently one can't control
the WebRTC IP handling policy, so this approach
is not available on Android.

Another thing is, of course, the fact that this speeds up
the process of launching an app.

And in general this approach seems more thorough to me.
It also provides another exfiltration security layer (a dummy proxy)
against regular HTTP requests as well,
i.e. it guards even against CSP bypass.

This uses the same approach that we took in the Tauri version:
https://github.com/deltachat/deltachat-desktop/pull/4852.

On top of that, the 500 limit itself is not something
that is set in stone and should be relied on.

To check that this still works, follow the instructions in
https://github.com/webxdc/webxdc-test/pull/40/files.