Improvements to Dialog Bottom Margin (#5494)
* remove dialog margin

* re-add dialog margin to settings. It is a seperate element, because that
way it is inside of the scroll area, not taking away from it, only
visible when you scrolled to the end.

* reintroduce bottom space for gallery, again by adding the margin to the
scroll content and not the container

* add spacing to remaining dialogs. using br line break beacuse it works
just fine and is easy to do and understand. again, idea here is to add
it to the content and not outside of it for the case that the dialog is
scrollable.

* Update CHANGELOG.md