fix: remove unnecessary `callback()` in QR
With this invocation the callback gets invoked twice in total.
This doesn't seem intentional.
Unfortunately the commit history is messy (gigantic squashes),
so it's hard to trace the intent and the effect.

But anyways currently all that `callback` does,
according to the usage of `processQr(`, is closing the dialog,
so this should not affect behavior, at least not in a bad way.