fix: consistently close dialog on "fingerprint" QR
The opened confirmation dialog doesn't seem to do much.
For some reason it only controls whether the parent "Scan QR"
dialog gets closed (i.e. what `callback` does).
So let's remove this unnecessary piece of complexity.

The logic (and the
`The fingerprint of ${contact.displayName} is valid` string)
seem to originate from 5641aea0c48e10a1970e85b648cb09be5263489d,
but it seems that it was buggy, because `handleScanResult`
was supposed to accept `chatId`,
but it was invoked with `boolean` argument.