refactor: split up `ChatList`'s search sections
This should not affect behavior, besides performance.

This commit is mostly about moving the code around,
without much changes to the logic.
Which is easy to see with a diff viewer that recognizes code moves.

Regarding performance: this ensures that we do not send
the contact and messages search queries (with an empty query)
to the backend as soon as the chat list is rendered.
Now the search queries will only get sent when the query is not
empty, i.e. when we actually do need to show the search results.
This is thanks to the fact that the new "search reults" components
are unrendered initially.
Namely, previously we would always load _all_ the contacts
from the backend.

This makes switching between accounts faster.