fix: draft WebXDC name being 'Unknown App' rarely
Can happen when adding a WebXDC attachment to the draft.

Currently this bug only happens if for some reason
`MessageListAndComposer` re-renders while a WebXDC is being added.
The fact that the bug doesn't manifest itself every time
is because in `useRef` we mutate the `draftRef`
instead of properly setting the state,
so a re-render with `draftState.file` being a temp file
gets skipped.
This state only lasts until `saveAndRefetchDraft_` finishes.

This bug has been discovered when refactoring `useDraft`
to get rid of the `useRef`:
https://github.com/deltachat/deltachat-desktop/pull/5612.