improvement: remove some too verbose "Close" btns
This saves space and should reduce confusion.
A "Cancel" button is sometimes useful, but a "Close" button
is pretty rarely.
Also for the most part this brings the look closer to that of
Delta Chat Android, except for the "See Reactions" dialog.

Replace them with a header "cross" in most places.