refactor: lift composer text state up, `useDraft`
This should greatly simplify the code,
and allow for further simplifications.

This is not production-ready yet, because this removes
the throttle / debounce.
See the new TODO comments.

However, even without the debounce, the performance
doesn't seem too bad.
Even if it is, the React compiler should be able
to negate the effects of this commit.

This also fixes a minor bug with `showSendButton`.