fix: debounce instead of throttle to save draft
And also save the draft when switching chats, or minimizing the app.

This the third and final commit in a series dedicated to
refactoring and fixing up `useDraft`.
The commit that removed `throttle` is actually not this one
but the first one in the series,
"refactor: lift composer text state up".

The performance, based on Chromium profiling and looking at how long
`keypress` event handler takes, seems to be pretty much unaffected
compared to when we relied on `throttle` in `ComposerMessageInputProps`.
It's still ~4ms.
And otherwise it's possible that it will actually improve,
because we send the draft to the backend much more rarely now.
An indirect thing it's that the chat list has to re-render
and to be refetched much less often, because `ChatlistItemChanged`
fires when the draft gets saved.

Closes https://github.com/deltachat/deltachat-desktop/issues/5252.

This should also improve the situation with
https://github.com/deltachat/deltachat-desktop/issues/3586
(but I'm not completely sure that things
aren't going to get worse at first).

This is also the "proper" resolution of
https://github.com/deltachat/deltachat-desktop/issues/3733,
which has previously been resolved by
d1fbc7ac3b16b607504a333d36f6cf51180cdae3
(https://github.com/deltachat/deltachat-desktop/pull/4144).
So this commit also somewhat reverts that one.