fix: remove `:emoji:` replacement on incoming msgs
Prior to this, we would render a message saying
`Foo :smile: bar` as `Foo 😄 bar`.
While this might be still needed for older messages
that use this format, it's hard to imagine
that anyone is actually using this format nowadays,
so let's render incoming messages as is in this regard.

We still do this replacement when sending messages,
ever since 20ea65d29847459e9179f1f57a8048447cdadbea
(https://github.com/deltachat/deltachat-desktop/pull/1781),
and it's probably enough (given that we need to do that at all).

This should improve performance a little, and simplify the code.