rename 'Save' to 'Save Message'
'Save' was sometimes brought into relation with 'File / Save',
esp. if one is used to that old menu ordering.

this PR tries to mitigate that confusion by not using the term 'Save',
but the more explicit 'Save Message'.

let's see if that mitigates the confusion.

(it is known that there are other ideas around,
that, however, require more effort and/or may add other confusions :)