fix: "Replace Draft?" prompt not appearing
Namely, when doing
- 'Share Profile'
- `webxdc.sendToChat()`
- bot command click
- `mailto:...?body=...` click

The issue was described in
https://github.com/deltachat/deltachat-desktop/pull/5643.
The problem was that, since we save the draft the the backend
on debounce and not immediately, it is not enough
to utilize `BackendRemote.rpc.getDraft()` to check whether
a draft is present.

So let's instead treat the source of truth for the draft state
as being inside of `useDraft`, and overall move more draft logic
to inside of `useDraft`.

The problem with this, however, is that `useDraft` is not always
rendered with the `accountId` and `chatId` that we need,
so we have to store the "set draft request"
until the correct chat has been selected and loaded.

Co-authored-by: Nico de Haen <ndh@deltachat.de>