fix: some bugs when switching chats
While working on `useDraft`, I found myself trying to implement
complex logic to manage outdated async functions,
deciding when and when not to set state.

However, `key` is just much simpler and better.

We do not rely on any kind of cross-chat state,
so having `key={chat.id}` makes sense.

The performance for switching chats appears to be about the same.