refactor: remove unnecessary `clearDraftState`
It was introduced in 2b3cc19c50f4b0d7910de6777fc04d15fc5ce4b8
(https://github.com/deltachat/deltachat-desktop/pull/4394),
and the idea was to disable the `Ctrl + Up` shortcut
(that quotes a message):
https://github.com/deltachat/deltachat-desktop/issues/4389.

However, this complicates the code unnecessarily.
Loading the draft is still possible, so let's load it.

With this change, and the fact that `useDraft` usage is behind
React `key`, now the draft only gets reloaded when
`isContactRequest` (or `inputRef`) changes.