refactor: `useDraft`: rm `isContactRequest` dep
This should not affect behavior, because `isContactRequest`
is only expected to change once, and only for chats
where there is no draft in the first place.

This hook dependency was introduced in
e7a0ce73efd28e3737c3de8342c2e531eb785799.
The issue, as usual, was that the composer input
would remain disabled even after you have accepted
the contact request.

However, this was back when its `disabled` attribute was controlled
imperatively, with `ref.setText()` or `ref.setState()`
from inside of `useDraft`. Now `disabled` it's simply equal
to `draftIsLoading`, which is not gonna remain `true` indefinitely.

With this change, and the fact that `useDraft` usage is behind
React `key`, now the draft only gets loaded once.