fix: don't override draft after load if no draft
This affects the case where the draft has been modified by the user
before we have managed to finish the initial `getDraft`.
Previously we would always reset the draft
to the "empty draft" state, even if `getDraft` returned `null`
(no draft). Now we only override if there _is_ a draft stored.

Although currently the composer textarea is disabled
while the draft is loading, it's still possible
to quote a message or add a file while the draft is loading,
so this is not just a refactor.